- Added missing @Override annotations.  - Fixed other programming issues reported by Eclipse.

+++ modules/com.noelios.restlet/src/com/noelios/restlet/util/CookieUtils.java
@@ -132,7 +132,7 @@ public class CookieUtils {
             if (maxAge >= 0) {
                 if (version == 0) {
                     long currentTime = System.currentTimeMillis();
-                    long maxTime = ((long) maxAge * 1000L);
+                    long maxTime = (maxAge * 1000L);
                     long expiresTime = currentTime + maxTime;
                     Date expires = new Date(expiresTime);
                     destination.append("; Expires=");
